SDR-101 feat: App에서 로그인 유무에 따라 라우팅 분기 처리
