+++ fe/src/pages/Login/Login.tsx
@@ -3,38 +3,12 @@ import TEXT from 'constants/text';
 import Logo from 'common/Logo/Logo';
 import LoginButton from 'components/Login/LoginButton/LoginButton';
 import LoginInput from 'components/Login/LoginInput/LoginInput';
-import { useUserInfo } from 'context/userInfoProvider';
-import { useEffect } from 'react';
-import { Navigate } from 'react-router-dom';
-import { fetchDataThatNeedToLogin } from 'utils/utils';
 import { Form, KakaoLogin, Wrap } from './Login.styled';
 
 const { KAKAO, NORMAL } = TEXT.LOGIN_BTN;
 const { ID, PASSWORD } = TEXT.INPUT;
 
 function Login() {
-  const [, dispatchUserInfo] = useUserInfo();
-  const accessToken = localStorage.getItem('accessToken');
-  const hasAccessToken = accessToken && accessToken !== 'null';
-
-  useEffect(() => {
-    if (hasAccessToken) {
-      setMyInfo();
-    }
-
-    async function setMyInfo() {
-      const myInfo = await fetchDataThatNeedToLogin(`${DOMAIN}/api/users/me`);
-      const { id, nickname, profileImage } = myInfo.data;
-      const myInfoJson = JSON.stringify({ userId: id, nickname, profileImageUrl: profileImage });
-      localStorage.setItem('MY_INFO', myInfoJson);
-      dispatchUserInfo({ type: 'SET_USER', userId: id, nickname, profileImageUrl: profileImage });
-    }
-  }, []);
-
-  if (hasAccessToken) {
-    return <Navigate to="/reviewList" />;
-  }
-
   return (
     <Wrap>
       <Form>
